requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
google-api-python-client==2.86.0
transformers>=4.31.0
torch>=2.0.0
beautifulsoup4>=4.9.3
requests>=2.25.1
